emacs initializong without errors on fedora 23

+++ setup.sh
@@ -4,8 +4,11 @@
 REPODIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 cd ~
-rm -irf .emacs.d	
-ln -s $REPODIR/.emacs.d
-ln -s $REPODIR/.gnus
-mv .emacs .emacs.bak
-
+rm -irf .emacs.d
+ln -s $REPODIR/emacs.d .emacs.d  > /dev/null 2>&1
+ln -s $REPODIR/gnus .gnus  > /dev/null 2>&1
+if [[ ! -f ~/.emacs.bak ]]; then
+  if [[ -f .emacs ]]; then
+    mv .emacs .emacs.bak
+  fi
+fi
